evaluation/pedestrian/nonparametric-hmc/requirements.txt
numpy==1.26.4
torch==1.13.1

pyro-ppl==1.4.0

matplotlib>=3.4.1
scipy>=1.6.3

pandas
seaborn
tqdm
matplotlib
